requirements.txt
hydra-core~=1.1
matplotlib~=3.5
numpy~=1.22
omegaconf~=2.1
scikit_learn~=1.0
seaborn~=0.11
setuptools~=60.5
torch~=1.10
torchvision~=0.11
